+++ bridge/doctrine-phpcr-odm/tests/Functional/CollectionType/ObjectUnrestrictTest.php
@@ -94,22 +94,4 @@ class ObjectUnrestrictTest extends PhpcrOdmTestCase
 
         $this->documentManager->find(null, '/test/article');
     }
-
-    private function createArticleSlideshow()
-    {
-        $article = new Article();
-        $article->id = '/test/article';
-
-        $image1 = $this->createImage('/path/to/image1', 100, 200, 'image/jpeg');
-        $image2 = $this->createImage('/path/to/image2', 100, 200, 'image/jpeg');
-        $image3 = $this->createImage('/path/to/image3', 100, 200, 'image/jpeg');
-
-        $article->slideshow = [
-            $image1,
-            $image2,
-            $image3,
-        ];
-
-        return $article;
-    }
 }
