+++ bridge/doctrine-phpcr-odm/tests/Functional/mappings/Psi.Component.ContentType.Tests.Functional.Example.Model.Article.dcm.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<doctrine-mapping
+    xmlns="http://doctrine-project.org/schemas/phpcr-odm/phpcr-mapping"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://doctrine-project.org/schemas/phpcr-odm/phpcr-mapping
+    https://github.com/doctrine/phpcr-odm/raw/master/doctrine-phpcr-odm-mapping.xsd"
+    >
+
+    <document 
+        name="Psi\Component\ContentType\Tests\Functional\Example\Model\Article"
+        referenceable="true"
+        >
+        <id name="id" />
+    </document>
+
+</doctrine-mapping>
